Rename context file.
